Create explicit empty and null values for some types
IMO these are different than f.e. the Default impls on TxOut etc,
because these types actually in many actual production cases want to be
set to the empty or none value. As opposed to Default on other types
that only exist to aid in test scenarios or so to avoid having to set
the fields.