remove --all from cargo test
according to the help the option is deprecated and equivalent to `--workspace`
and we haven't a workspace.
Moreover, and most importantly, the option somehow activates features which
should not be present. for example:
`cargo tree --no-default-features -e no-dev --all | grep serde` has dependencies in the output while without `--all` returns nothing