pset: breaking, encode witness with TxOutWithWitness
TxOut is encoded/decoded without the witness, it is handled separetely during
transaction encoding/decoding (segwit).
For PSET the TxOut encoding is reused but in this case the witness is not
handled separately.
This commit introduce TxOutWithWitness which is logically a TxOut with the
exception that also the witness is encoded/decoded.
4 Existing tests breaks with this change because they are using string test
vectors not containing the witness. To avoid too many changes this tests
are temporary ignored, they will be re-added in following commit in the same PR.