pset: restore 2 tests
pset_swap_tutorial test vector is updated with the updated serialization of
TxOutWithWitness.

Note the old test vector doesn't contain a input.witness_utxo.witness but the
changed encoding require to find an empty value there.

To achieve so, decode of TxOutWitness has been temporary changed to skip
the witness decoding, the parsed PSET is then encoded with the updated encoding.
The result is then used as `pset_swap_tutorial.hex`