bech32: replace blech32 implementation
Essentially, we copied the src/primitives/decode.rs file from the bech32
crate, which provides some convenience types for parsing HRP strings
that have witness versions and where the checksum is chosen selectively
based on which witness version.

This is stupid BIP-173-specific behavior so I don't think it makes sense
to try to generalize rust-bech32 to enable this. Copying the file is
fine.