transaction: range-check pegin data when parsing
It is possible to cause a panic by calling `TxIn::pegin_data` on a
malformed pegin, which has 6 entries and the last one (which should be a
Merkle proof of block inclusion) is too small to even have a blockhash
in it.

On the actual chain this should not be possible because invalid pegin
data would be rejected.