add Address::is_liquid
User code can already compare against AddressParams, but it seems more
easy to find a method to check if the address is liquid.
Even thoug is less general because you can't discriminate between
elements and liquid testnet, most of the time wallets are interested
only in the mainnet or not discrimination.