clippy: fix if_not_else lint
This one is a bit silly -- it complains if you do `if !x {} else {}`
because it would supposedly be simpler to swap the if and the else.

In fact, everywhere it triggers in this crate it *would* be an
improvement, or at least neutral. So turn it on. In some cases we are
using if-statements where we could instead use matches.