locktime: use bitcoin-units locktime types
In bitcoin-units 1.0 we have types that represent locktime times and
heights. The locktimes themselves are not in units (they will be exposed
in bitcoin-primitives once that comes out) but the essential hard logic
is implemented on these supporting types.

Importantly, the new types don't implement serde, which is why we needed
to do breaking serde changes in the last couple of commits. (But we
should have done these changes a long time ago, anyway.)