locktime: replace serde impl with u32-based one
This completely changes the serde format for `LockTime`, which in turn
changes the format for `Transaction`. This will break e.g. hal-elements.
However, the existing format is kinda ridiculous and I don't believe it
is used anywhere.

Alarmingly this change does not break a single test. We should add some
regression tests.