pset: replace ad-hoc serde implementation with base64 string
There is a standard serialization of PSETs (and PSBTs, and PSBT2s): the
base64 of the byte encoding. Rather than making up our own serialization
based on the details of our rust-elements types, which makes basically
every structural property of our types a property of our API, just use
the standard encoding.

This is a major breaking change; we may need a compatibility layer based
on rust-elements 0.25.