transaction and block: use Weight as return type for weight methods
The API change is actually only a couple LOC. The real change is that we
start using u64 arithmetic to compute weights rather than usize
arithmetic, because Weight::from_wu wants a u64.

To facilitate this we introduce the private Len64 trait which provides
lengths in u64, and use it throughout the crate. This replaces 23 casts
to usize or u64 with four.