encode: implement Encodable and Decodable for Vec<T>
Right now we have explicit impls of Vec<T> for a ton of specific types T.

This is annoying for users, who are unable to implement Vec<T> themselves
for their own types. In particular is super annoying for me trying to
do a semver trick with elements 0.25 and 0.26, since I can't reexport
the 0.26 Encodable trait since I would need to implement it for Vec<TxOut>
for the 0.25 TxOut type. (But conversely, if I retain the full trait
definition in 0.25, I can't blanket-impl this trait for all T: Encodable
because of this Vec<T> crap; so I wind up needing to explicitly implement
Encodable and Decodeable for every single type I re-export, and if I
forget any then that's a semver violation. Brutal.)

We do this is because we need a specialized impl for Vec<u8> and Rust has
no specialization. However, Kixunil pointed out on rust-bitcoin that we
can just do "runtime specialization" by checking TypeId::of::<T> and doing
the efficient thing when T is u8.

In practice, because the compiler knows T at codegen time, it will know
how the check goes and simply delete the check and the untaken branch,
so we have exactly the same result as we would if we had specialization.

This works only for T: 'static, but it turns out that all of our types
that we currently impl Encodable for Vec<T> on are 'static (and it's
hard to imagine one that wouldn't be, since all bitcoin/elements primitive
types are simple PODs).

It also seems plausible that a user could construct a Vec<u8> which the
compiler somehow doesn't realize is a Vec<u8> (e.g. if they had a
Vec<dyn Encodable> or something, in some future Rust version that
supports that). Such code will be sound and correct, but be slow because
it won't use the fast path. But I don't think this is a practical
concern because I can't even figure out how to do it..